Add files via upload
Calculates the triangular number of any given n for x number of times. Uses scanf() function, if to exit the program, and a nested for loop statement to create the correct output. A book chapter gave me partial code only giving a simple program with no loops or if statement. Attempting to use existing knowledge to change and improve on it. 

Source: Programming in C 4th Edition/Kochan